Update db migration file

+++ TweetService/db/migrations/001_create_tweets.up.sql
@@ -2,7 +2,7 @@
 
 Create table tweets (
 	id serial primary key,
-    user_id bigint unsigned not null,
+    user_id int not null,
     tweet varchar(280) not null,
     createdAt timestamp default current_timestamp,
     updatedAt timestamp default current_timestamp
